refactor(features): moved voteNow to features
