+++ src/components/VotingCandidate/index.tsx
@@ -3,7 +3,7 @@ import ArrowUpwardIcon from '@mui/icons-material/ArrowUpward';
 import CloseIcon from '@mui/icons-material/Close';
 import CandidatePosition from 'styled/CandidatePosition';
 import getLocaleDate from 'utils/helperFunctions/getLocaleDate';
-import voteNow from 'utils/helperFunctions/voteNow';
+import voteNow from 'features/voteNow';
 import useCandidateVotes from 'hooks/useCandidateVotes';
 import useUserRemainingVotes from 'hooks/useUserRemainingVotes';
